gun: merge Gun_FireWeapon
