+++ src/libtrx/include/libtrx/game/spawn.h
@@ -1,9 +1,12 @@
 #pragma once
 
 #include "./items/types.h"
+#include "./types.h"
 
 void Spawn_Splash(const ITEM *item);
 
 extern int16_t Spawn_Blood(
     int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
     int16_t room_num);
+
+extern void Spawn_Ricochet(const GAME_VECTOR *pos);
