+++ src/tr2/game/gun/gun_misc.c
@@ -201,103 +201,6 @@ void Gun_ChangeTarget(const WEAPON_INFO *const weapon)
     Gun_TargetInfo(weapon);
 }
 
-int32_t Gun_FireWeapon(
-    const LARA_GUN_TYPE weapon_type, ITEM *const target, const ITEM *const src,
-    const int16_t *const angles)
-{
-    const WEAPON_INFO *const weapon = &g_Weapons[weapon_type];
-    const LARA_INFO *const lara = Lara_GetLaraInfo();
-    AMMO_INFO *const ammo = Gun_GetAmmoInfo(weapon_type);
-    ASSERT(ammo != nullptr);
-
-    if (ammo == &lara->pistol_ammo || Game_IsBonusFlagSet(GBF_NGPLUS)) {
-        ammo->ammo = 1000;
-    }
-
-    if (ammo->ammo <= 0) {
-        ammo->ammo = 0;
-        return 0;
-    }
-
-    ammo->ammo--;
-
-    const XYZ_32 view_pos = {
-        .x = src->pos.x,
-        .y = src->pos.y - weapon->gun_height,
-        .z = src->pos.z,
-    };
-    const XYZ_16 view_rot = {
-        .x = angles[1]
-            + weapon->shot_accuracy * (Random_GetControl() - DEG_90) / DEG_360,
-        .y = angles[0]
-            + weapon->shot_accuracy * (Random_GetControl() - DEG_90) / DEG_360,
-        .z = 0,
-    };
-    Matrix_GenerateW2V(&view_pos, &view_rot);
-
-    SPHERE spheres[33];
-    int32_t sphere_count = Collide_GetSpheres(target, spheres, false);
-    int32_t best_sphere = -1;
-    int32_t best_dist = 0x7FFFFFFF;
-
-    for (int32_t i = 0; i < sphere_count; i++) {
-        const SPHERE *const sphere = &spheres[i];
-        const int32_t r = sphere->r;
-        if (ABS(sphere->pos.x) < r && ABS(sphere->pos.y) < r
-            && sphere->pos.z > r
-            && SQUARE(sphere->pos.x) + SQUARE(sphere->pos.y) <= SQUARE(r)) {
-            const int32_t dist = sphere->pos.z - r;
-            if (dist < best_dist) {
-                best_dist = dist;
-                best_sphere = i;
-            }
-        }
-    }
-
-    Stats_AddAmmoUsed();
-
-    GAME_VECTOR start;
-    start.pos.x = view_pos.x;
-    start.pos.y = view_pos.y;
-    start.pos.z = view_pos.z;
-    start.room_num = src->room_num;
-
-    if (best_sphere < 0) {
-        const int32_t dist = weapon->target_dist;
-        GAME_VECTOR hit_pos;
-        hit_pos.x = view_pos.x + ((dist * g_MatrixPtr->_20) >> W2V_SHIFT);
-        hit_pos.y = view_pos.y + ((dist * g_MatrixPtr->_21) >> W2V_SHIFT);
-        hit_pos.z = view_pos.z + ((dist * g_MatrixPtr->_22) >> W2V_SHIFT);
-        hit_pos.room_num = src->room_num;
-        Room_GetSector(hit_pos.x, hit_pos.y, hit_pos.z, &hit_pos.room_num);
-        const bool object_on_los = LOS_Check(&start, &hit_pos);
-        if (Gun_SmashItems(start.pos, hit_pos.pos, &hit_pos.pos)
-            == PROJECTILE_HIT_STOP) {
-            Room_GetSector(hit_pos.x, hit_pos.y, hit_pos.z, &hit_pos.room_num);
-        }
-        if (!object_on_los) {
-            Spawn_Ricochet(&hit_pos);
-        }
-        return -1;
-    } else {
-        Stats_AddAmmoHits();
-        GAME_VECTOR hit_pos;
-        hit_pos.pos.x =
-            view_pos.x + ((best_dist * g_MatrixPtr->_20) >> W2V_SHIFT);
-        hit_pos.pos.y =
-            view_pos.y + ((best_dist * g_MatrixPtr->_21) >> W2V_SHIFT);
-        hit_pos.pos.z =
-            view_pos.z + ((best_dist * g_MatrixPtr->_22) >> W2V_SHIFT);
-        hit_pos.room_num =
-            Room_GetIndexFromPos(hit_pos.pos.x, hit_pos.pos.y, hit_pos.pos.z);
-        Gun_SmashItems(start.pos, hit_pos.pos, nullptr);
-        Gun_HitTarget(
-            target, &hit_pos,
-            weapon->damage * (Game_IsBonusFlagSet(GBF_JAPANESE) ? 2 : 1));
-        return 1;
-    }
-}
-
 PROJECTILE_HIT Gun_SmashItems(
     const XYZ_32 start, const XYZ_32 target, XYZ_32 *const out_hit_pos)
 {
