+++ src/tr2/game/spawn.h
@@ -31,5 +31,3 @@ int16_t Spawn_Knife(
 void Spawn_BloodBath(
     int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
     int16_t room_num, int32_t count);
-
-void Spawn_Ricochet(const GAME_VECTOR *pos);
